season policy and season test is compelted

+++ modules/Hadihosseini88/Course/Http/Controllers/SeasonController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Hadihosseini88\Common\Responses\AjaxResponses;
 use Hadihosseini88\Course\Http\Requests\SeasonRequest;
 use Hadihosseini88\Course\Models\Season;
+use Hadihosseini88\Course\Repositories\CourseRepo;
 use Hadihosseini88\Course\Repositories\SeasonRepo;
 
 class SeasonController extends Controller
@@ -18,10 +19,10 @@ class SeasonController extends Controller
         $this->seasonRepo = $seasonRepo;
     }
 
-    public function store($course,SeasonRequest $request)
+    public function store($course,SeasonRequest $request, CourseRepo $courseRepo)
     {
+        $this->authorize('createSeason', $courseRepo->findByid($course));
         $this->seasonRepo->store($course, $request);
-
         newFeedback();
         return back();
     }
@@ -29,11 +30,13 @@ class SeasonController extends Controller
     public function edit($id)
     {
         $season = $this->seasonRepo->findByid($id);
+        $this->authorize('edit', $season);
         return view('Courses::seasons.edit', compact('season'));
     }
 
     public function update($id, SeasonRequest $request)
     {
+        $this->authorize('edit', $this->seasonRepo->findByid($id));
         $this->seasonRepo->update($id, $request);
         newFeedback();
         return back();
@@ -42,6 +45,7 @@ class SeasonController extends Controller
     public function destroy($id)
     {
         $season = $this->seasonRepo->findByid($id);
+        $this->authorize('delete', $season);
         $season->delete();
 
 //        return AjaxResponses::SuccessResponse();
@@ -49,14 +53,14 @@ class SeasonController extends Controller
 
     public function accept($id, SeasonRepo $seasonRepo)
     {
-//        $this->authorize('change_confirmation_status',Season::class);
+        $this->authorize('change_confirmation_status',Season::class);
         if ($seasonRepo->updateConfirmationStatus($id,Season::CONFIRMATION_STATUS_ACCEPTED)){
             return AjaxResponses::SuccessResponse();
         }
         return AjaxResponses::FailedResponse();
     }
     public function reject($id, SeasonRepo $seasonRepo) {
-//        $this->authorize('change_confirmation_status',Season::class);
+        $this->authorize('change_confirmation_status',Season::class);
         if ($seasonRepo->updateConfirmationStatus($id,Season::CONFIRMATION_STATUS_REJECTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -65,7 +69,7 @@ class SeasonController extends Controller
 
     public function lock($id, SeasonRepo $seasonRepo) {
 
-//        $this->authorize('change_confirmation_status',Season::class);
+        $this->authorize('change_confirmation_status',Season::class);
         if ($seasonRepo->updateStatus($id,Season::STATUS_LOCKED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -73,7 +77,7 @@ class SeasonController extends Controller
     }
 
     public function unlock($id, SeasonRepo $seasonRepo) {
-//        $this->authorize('change_confirmation_status',Season::class);
+        $this->authorize('change_confirmation_status',Season::class);
         if ($seasonRepo->updateStatus($id,Season::STATUS_OPENED)){
             return AjaxResponses::SuccessResponse();
         }
